tile-server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openslide-python>=1.3.0
Pillow>=10.0.0
